Allow specifying custom converters in Decoder
Add a new Decoder implementation, CustomDecoder, which allows users to specify custom
type converters via its static factory method, create. To avoid duplicate code, introduce
a new package private base class, AbstractRegistryDecoder, which the existing DefaultDecoder
and CustomDecoder both extend. CustomDecoder will add the default factories used by
DefaultDecoder as well, placing them after the custom ones so that they may be overridden
and are only used as fallbacks.
